kafka producer implementation
